fix : input logics
